+++ Inventory/src/main/java/com/bartolay/inventory/repositories/SupplierRepository.java
@@ -6,12 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bartolay.inventory.entity.Supplier;
 
 @Repository
-public interface SupplierRepository extends CrudRepository<Supplier, Long>{
+public interface SupplierRepository extends CrudRepository<Supplier, Integer>{
     
-//    public ObservableList<Supplier> getSuppliers();
-//    public Supplier getSupplier(long id);
-//    public void saveSuplier(Supplier supplier);
-//    public void updateSuplier(Supplier supplier);
-//    public void deleteSuplier(Supplier supplier);
-//    public ObservableList<String> getNames();
 }
